LayoutTests/fast/parser/entity-end-style-tag.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<style id="mystyle">&lt;/style></style>
</head>
<body>
<script>
description("Test parsing of entity-escaped &lt;/style&gt; tag for <a href=\"https://bugs.webkit.org/show_bug.cgi?id=7931\">Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely</a>");

var expectedResult = '&lt;' + '/style>';
shouldBe('document.getElementById("mystyle").textContent', 'expectedResult');
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
