third_party/blink/web_tests/fast/parser/entity-end-style-tag-expected.txt

Test parsing of entity-escaped </style> tag for Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("mystyle").textContent is expectedResult
PASS successfullyParsed is true

TEST COMPLETE
